split test files.
