+++ tests/test_builder.py
@@ -70,94 +70,3 @@ def test_laziness():
 def test_raw():
     assert Builder().literally('foo').raw('b[a-z]r').is_valid()
 
-def test_parse():
-    assert SRL('literally "abc"').match('abc')
-    assert SRL('one of "abc"').match('a')
-    assert SRL('letter').match('a')
-    assert SRL('letter from a to c').match('b')
-    assert not SRL('letter from a to c').match('d')
-    assert SRL('uppercase letter').match('A')
-    assert SRL('letter from A to C').match('B')
-    assert not SRL('letter from A to C').match('D')
-    assert SRL('any character').match('a')
-    assert SRL('no character').match(' ')
-    assert SRL('digit').match('0')
-    assert not SRL('anything').match('\n')
-    assert SRL('anything').match('-')
-    assert SRL('new line').match('\n')
-    assert SRL('whitespace').match(' ')
-    assert SRL('no whitespace').match('a')
-    assert SRL('tab').match('\t')
-    assert SRL('raw "[a-zA-Z]"').match('a')
-    assert not SRL('raw "[a-z]"').match('A')
-    assert SRL('digit exactly 2 times').match('00')
-    assert SRL('digit exactly 2 times').match('000').group() == '00'
-    assert not SRL('digit between 2 and 3 times').match('1')
-    assert SRL('digit between 2 and 3 times').match('12')
-    assert SRL('digit between 2 and 3').match('12')
-    assert SRL('digit optional').match('a')
-    assert not SRL('digit once or more').match('a')
-    assert SRL('digit once or more').match('1')
-    assert SRL('digit once or more').match('12')
-    assert SRL('digit never or more').match('a')
-    assert SRL('digit never or more').match('a1')
-    assert not SRL('digit at least 2 times').match('1')
-    assert SRL('digit at least 2 times').match('11')
-    assert SRL('capture "abc"').match('abc')
-    assert SRL('capture (digit once or more)').match('1').groups() == ('1', )
-    assert SRL('capture (digit once or more) as "mydigit"').match('1').groups() == ('1', )
-    assert SRL('any of (literally "sample")').match('sample')
-    assert SRL('capture (anything once or more) until "m"').match('example')
-    assert SRL('begin with must end').match('')
-    assert SRL('letter case insensitive').match('A')
-    assert SRL('capture (letter once or more) all lazy').match('a')
-
-def test_rules():
-    for filename in glob('tests/rules/*.rule'):
-        with open(filename) as f:
-            fdata = f.read()
-            in_capture = False
-            data = {'srl': None, 'matches': [], 'no_matches': [], 'captures': {}}
-            for line in fdata.splitlines():
-                if not line or line.startswith('#'):
-                    continue
-                if in_capture and not line.startswith('-'):
-                    in_capture = False
-                if line.startswith('srl: '):
-                    data['srl'] = line[5:]
-                elif line.startswith('match: "'):
-                    data['matches'].append(line[8:-1])
-                elif line.startswith('no match: "'):
-                    data['no_matches'].append(line[11:-1])
-                elif line.startswith('capture for "') and line.endswith('":'):
-                    in_capture = line[13:-2]
-                elif in_capture and line.startswith('-'):
-                    split = line[1:].split(': ')
-                    index = int(split[0].strip())
-                    name = split[1].strip()
-                    matching = split[2][1:-1]
-                    data['captures'].setdefault(in_capture, {})
-                    if name == '0':
-                        data['captures'][in_capture].setdefault(index, [])
-                        data['captures'][in_capture][index].append(matching)
-                    else:
-                        data['captures'][in_capture].setdefault(index, {})
-                        data['captures'][in_capture][index][name] = matching
-            srl = SRL(data['srl'])
-            for match in data['matches']:
-                assert srl.search(match), match
-            for no_match in data['no_matches']:
-                assert not srl.search(no_match), no_match
-            for capture, matches in data['captures'].items():
-                for index, expected in matches.items():
-                    if isinstance(expected, list):
-                        assert expected[0] == srl.findall(capture)[index]
-                    else:
-                        groupindex = srl.groupindex
-                        all = srl.findall(capture)
-                        for key, val in expected.items():
-                            find_val = srl.findall(capture)[index]
-                            if isinstance(find_val, str):
-                                assert val == find_val
-                            else:
-                                assert val == find_val[groupindex[key] - 1]
